Set up global world update
