+++ core/client_side/game_mechanics/logic/game_components/ball.js
@@ -1,0 +1,45 @@
+
+import * as math from '../../../../_lib/math';
+import * as shapes from '../geometry_shapes';
+import {SolidBody} from '../solid_body';
+
+
+export class Ball extends SolidBody {
+    constructor(radius) {
+        super();
+        this._circle = new shapes.Circle(radius);
+    }
+
+    get radius() {
+        return this._circle.radius;
+    }
+
+    bounce(normVec) {
+        let normVec0 = math.divide(normVec, math.norm(normVec));
+
+        let normMatrix = [
+            [normVec0[0] * normVec0[0], normVec0[0] * normVec0[1]],
+            [normVec0[0] * normVec0[1], normVec0[1] * normVec0[1]]
+        ];
+
+        let identityMatrix = math.eye(2);
+
+        let transformMatrix = math.subtract(identityMatrix, math.multiply(normMatrix, 2));
+
+        let newVelocity = math.multiply(transformMatrix, this.velocity).toArray();
+        newVelocity.forEach(x => -x);
+
+        this.velocity = newVelocity;
+    }
+
+    draw(canvas) {
+        let context = canvas.getContext("2d");
+        let position = this.position;
+        context.beginPath();
+        context.arc(position[0], position[1], this.radius, 0, 2 * Math.PI, false);
+        context.fillStyle = 'green';
+        context.fill();
+        context.lineWidth = 1;
+        context.stroke();
+    }
+}
