+++ core/client_side/game_mechanics/main/game_start.js
@@ -2,9 +2,9 @@
 import {Game} from './game';
 
 import {WSEndpoint} from '../network/endpoint'  //TODO remove
-import {ClientServerDispatcher} from '../network/game_event_dispatcher'   //TODO remove
+import {GameEventDispatcher} from '../network/game_event_dispatcher'   //TODO remove
 window.wse = new WSEndpoint();    //TODO remove
-window.disp = new ClientServerDispatcher(); //TODO remove
+window.disp = new GameEventDispatcher(); //TODO remove
 
 
 //const CANVAS_ID = "game";
